adding licence file + build tags
