+++ worker/config.go
@@ -72,6 +72,7 @@ type Options struct {
 	CacheMb int64
 
 	AuditEnabled bool
+	AuditDir string
 }
 
 // Config holds an instance of the server options..
@@ -96,12 +97,12 @@ func (opt *Options) validate() {
 	x.Check(err)
 	td, err := filepath.Abs(x.WorkerConfig.TmpDir)
 	x.Check(err)
-	ad, err := filepath.Abs(x.WorkerConfig.AuditDir)
+	ad, err := filepath.Abs(opt.AuditDir)
 	x.Check(err)
 	x.AssertTruef(pd != wd, "Posting and WAL directory cannot be the same ('%s').", opt.PostingDir)
 	x.AssertTruef(pd != td, "Posting and Tmp directory cannot be the same ('%s').", opt.PostingDir)
 	x.AssertTruef(wd != td, "WAL and Tmp directory cannot be the same ('%s').", opt.WALDir)
-	x.AssertTruef(ad != pd, "Posting and Audit Directory cannot be the same ('%s').", x.WorkerConfig.AuditDir)
-	x.AssertTruef(ad != wd, "WAL and Audit directory cannot be the same ('%s').", x.WorkerConfig.AuditDir)
-	x.AssertTruef(ad != td, "Tmp and Audit directory cannot be the same ('%s').", x.WorkerConfig.AuditDir)
+	x.AssertTruef(ad != pd, "Posting and Audit Directory cannot be the same ('%s').", opt.AuditDir)
+	x.AssertTruef(ad != wd, "WAL and Audit directory cannot be the same ('%s').", opt.AuditDir)
+	x.AssertTruef(ad != td, "Tmp and Audit directory cannot be the same ('%s').", opt.AuditDir)
 }
