+++ x/config.go
@@ -112,8 +112,6 @@ type WorkerOptions struct {
 	LogRequest int32
 	// If true, we should call msync or fsync after every write to survive hard reboots.
 	HardSync bool
-
-	AuditDir string
 }
 
 // WorkerConfig stores the global instance of the worker package's options.
